Level Design
Implemented a TON of assets of the Level Design. Got a temporary lobby setup above the map so you can look down over the edge at the map. (will probably need to spice up the lobby more tho as it's pretty plain rn)
Tweaked the airplane code a tiny bit, will need to test and keep tweaking if it's still stuttery